🧹 move WalkGlob into fsutil
